+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### unreleased
 - **fix:** flaws
+- **fix:** errors line breaks
 
 ### [v0.1.2] - 2016-06-25
 - **fix:** fix flaws